ddr_system/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
python-docx>=1.1.0
reportlab>=4.2.0
PyPDF2>=3.0.0
requests>=2.31.0
python-dotenv>=1.0.0
